+++ pyproject.toml
@@ -18,6 +18,7 @@ typing-extensions = "^4.12.2"
 
 [tool.poetry.dev-dependencies]
 pytest = "^8.3.1"
+python-dotenv = "^1.0.1"
 
 [tool.poetry.group.dev.dependencies]
 python-semantic-release = ">=8.3,<10.0"
