+++ realtime/channel.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 import asyncio
 import json
+import logging
 from typing import TYPE_CHECKING, Any, Dict, List, NamedTuple
 
 from realtime.message import ChannelEvents
-from realtime.push import Push
 from realtime.types import Callback
 
 from .presence import RealtimePresence
@@ -21,6 +21,35 @@ class CallbackListener(NamedTuple):
     event: str
     on_params: Dict[str, Any]
     callback: Callback
+
+
+class Push:
+    def __init__(self, channel: Channel, event: str, payload: Dict[str, Any] = {}):
+        self.channel = channel
+        self.event = event
+        self.payload = payload
+        self.ref = ""
+
+    def send(self):
+        asyncio.get_event_loop().run_until_complete(self._send())
+
+    async def _send(self):
+        self.ref = self.channel.socket._make_ref()
+
+        message = {
+            "topic": self.channel.topic,
+            "event": self.event,
+            "payload": self.payload,
+            "ref": self.ref,
+        }
+
+        try:
+            await self.socket.ws_connection.send(json.dumps(message))
+        except Exception as e:
+            logging.error(f"send push failed: {e}")
+
+    def update_payload(self, payload: Dict[str, Any]):
+        self.payload = {**self.payload, **payload}
 
 
 class Channel:
